1. add mediafile play(test on x86)

+++ media_engine/media_engine.py
@@ -1,5 +1,6 @@
 import enum
 import os
+import platform
 import signal
 import subprocess
 import time
@@ -7,12 +8,14 @@ import time
 import numpy as np
 from PyQt5.QtCore import QObject, pyqtSignal, QMutex, QThread
 
+import utils.utils_file_access
 from ext_qt_widgets.playing_preview_widget import PlayingPreviewWindow
 from media_engine.media_engine_def import PlayStatus, RepeatOption
 from utils.utils_ffmpy import get_ffmpeg_cmd_with_playing_media_file_
 from utils.utils_file_access import get_led_config_from_file_uri, get_int_led_config_from_file_uri
+from utils.utils_system import get_eth_interface
 from media_configs.video_params import VideoParams
-from global_def import log
+from global_def import log, root_dir, ETH_DEV, SU_PWD
 import socket
 from multiprocessing import shared_memory
 from ctypes import *
@@ -40,7 +43,7 @@ class MediaEngine(QObject):
         self.shm_sem_read_uri = "/shm_read_sem"
 
         os.system("pkill -f show_ffmpeg_shared_memory")
-        
+
         self.play_single_file_thread = None
         self.play_single_file_worker = None
 
@@ -122,6 +125,8 @@ class PlaySingleFileWorker(QObject):
 
     def __init__(self, media_engine: MediaEngine, file_uri, with_audio: bool, with_preview: bool):
         super().__init__()
+        self.shm = None
+        self.agent_process = None
         self.media_engine = media_engine
         self.file_uri = file_uri
         self.play_with_audio = with_audio
@@ -156,10 +161,17 @@ class PlaySingleFileWorker(QObject):
         self.pysignal_single_file_play_status_change.emit(self.play_status, self.playing_source)
 
     def run(self):
-        try_wait_write_flag = 0
-        os.system("pkill -f show_ffmpeg_shared_memory")
-        time.sleep(0.02)
-        # os.system("show_ffmpeg_shared_memory > /dev/null &")
+        try_wain_write_count_max = 5000
+        if platform.machine() in ('arm', 'arm64', 'aarch64'):
+            try_wain_write_count_max = 500
+        try_wait_write_count = 0
+        subprocess.Popen("pkill -f show_ffmpeg_shared_memory", shell=True)
+        # kill_agent_cmd = "sudo -S pkill -f show_ffmpeg_shared_memory"
+        # su_kill_agent_cmd = 'echo {} | '.format(SU_PWD) + kill_agent_cmd
+        # log.debug("su_kill_agent_cmd : %s", kill_agent_cmd)
+        # os.system(su_kill_agent_cmd)
+        time.sleep(1)
+
         while True:
             self.worker_status = 1
             if self.play_status != PlayStatus.Stop:
@@ -176,36 +188,37 @@ class PlaySingleFileWorker(QObject):
             self.shm = None
             while True:
                 try:
-                    #find agent preview window pos
+                    # find agent preview window pos
                     line = os.popen("xdpyinfo | awk '/dimensions/{print $2}'").read()
-
-                    print("lines : %s", line)
                     tmp = line.split("x")
-                    print("tmp[0] = ", tmp[0])
-                    print("tmp[1] = ", tmp[1])
                     geo_w = int(tmp[0])
                     geo_h = int(tmp[1])
-                    print("geo_w : ", geo_w)
-                    print("geo_h : ", geo_h)
 
-                    if self.output_width >= 1280 or self.output_height >=720:
+                    if self.output_width >= 1280 or self.output_height >= 720:
                         preview_pos_x = geo_w - 640
                         preview_pos_y = 320
                     else:
                         preview_pos_x = geo_w - self.output_width
                         preview_pos_y = 320
 
-                    #handle raw socket agent 
-                    os.environ['SDL_VIDEO_WINDOW_POS']="%d,%d" % (preview_pos_x,preview_pos_y)
-                    agent_cmd = "show_ffmpeg_shared_memory %d %d 1" % (self.output_width, self.output_height)
+                    # handle raw socket agent
+                    os.environ['SDL_VIDEO_WINDOW_POS'] = "%d,%d" % (preview_pos_x, preview_pos_y)
+                    ld_path = root_dir + "/ext_binaries"
+                    agent_cmd = (
+                            "%s/ext_binaries/show_ffmpeg_shared_memory %s %d %d 0 "
+                            % (root_dir, ETH_DEV, self.output_width, self.output_height))
+                    log.debug("agent_cmd : %s", agent_cmd)
                     self.agent_process = subprocess.Popen(agent_cmd, shell=True)
-                    log.debug("self.agent_process.pid = %d\n", self.agent_process.pid)
-                    self.shm = shared_memory.SharedMemory("posixsm", False, 0x400000);
+                    log.debug("self.agent_process.pid : %d", self.agent_process.pid)
+                    time.sleep(1)
+                    # self.agent_process = subprocess.Popen(agent_cmd, shell=True)
+                    # log.debug("self.agent_process.pid = %d\n", self.agent_process.pid)
+                    self.shm = shared_memory.SharedMemory("posixsm", False, 0x400000)
                 except Exception as e:
                     log.fatal(e)
 
-                    os.system("pkill -f show_ffmpeg_shared_memory")
-                    time.sleep(0.01)
+                    subprocess.Popen("pkill -f show_ffmpeg_shared_memory", shell=True)
+                    time.sleep(3)
 
                 if self.shm != None:
                     break
@@ -216,33 +229,37 @@ class PlaySingleFileWorker(QObject):
             if sem_write_flag == 0:
                 log.error("failed to create sem: %s", self.media_engine.shm_sem_write_uri)
                 return -1
-            
-            #Init the read sem
+
+            # Init the read sem
             sem_read_flag = self.shm_sem.sem_open(self.media_engine.shm_sem_read_uri, os.O_CREAT, 0x666, 0)
             if sem_read_flag == 0:
                 log.error("failed to create sem: %s", self.media_engine.shm_sem_read_uri)
                 return -1
-            
+
             ffmpeg_cmd = get_ffmpeg_cmd_with_playing_media_file_(self.file_uri,
                                                                  width=self.output_width, height=self.output_height,
                                                                  target_fps="24/1", image_period=20)
             log.debug("ffmpeg_cmd : %s", ffmpeg_cmd)
-            self.ff_process = subprocess.Popen(ffmpeg_cmd, stdout=subprocess.PIPE, bufsize=10 ** 8, shell=True)
+            try:
+                self.ff_process = subprocess.Popen(ffmpeg_cmd, stdout=subprocess.PIPE, bufsize=10 ** 8, shell=True)
+            except Exception as e:
+                log.error(e)
             # write__flag_tmp = self.shm_sem.sem_post(sem_write_flag)
             if self.ff_process.pid > 0:
                 self.play_status_change(PlayStatus.Playing, self.file_uri)
-                while True:
+                self.image_from_pipe = None
+                while self.ff_process.pid > 0:
                     self.image_from_pipe = self.ff_process.stdout.read(self.output_width * self.output_height * 3)
                     if len(self.image_from_pipe) <= 0:
                         log.debug("play end")
-                        self.agent_process.kill()
+                        if self.agent_process is not None:
+                            self.agent_process.kill()
                         break
 
                     if self.play_with_preview:
                         try:
                             self.raw_image = np.frombuffer(self.image_from_pipe, dtype='uint8')
                             self.raw_image = self.raw_image.reshape((self.output_height, self.output_width, 3))
-
 
                             self.ff_process.stdout.flush()
                             self.pysignal_refresh_image_pixmap.emit(self.raw_image)
@@ -254,39 +271,42 @@ class PlaySingleFileWorker(QObject):
                         if self.ff_process is not None:
                             os.kill(self.ff_process.pid, signal.SIGTERM)
                         break
-                    
-                    # test shm
-                    #print("w : %s" % time.time())
-                    #start_time = time.time()
-                    #print("typeof shm buf : %s\n", type(self.shm._buf))
-                    #print("typeof self.image_from_pipe : %s\n", type(self.image_from_pipe))
-                    
-                    #wait the write sem for write
-                    #log.debug("wait write sem!\n");
-                    #write_flag_tmp = self.shm_sem.sem_wait(sem_write_flag)
+
                     write_flag_tmp = self.shm_sem.sem_trywait(sem_write_flag)
                     if write_flag_tmp == -1:
-                        try_wait_write_flag += 1
-                        if try_wait_write_flag > 500:
-                            try_wait_write_flag = 0
+                        try_wait_write_count += 1
+                        if try_wait_write_count > 500:
+                            try_wait_write_count = 0
                             log.error("missing agent!")
-                            os.system("pkill -f show_ffmpeg_shared_memory")
-                            time.sleep(0.01)
-                            agent_cmd = "show_ffmpeg_shared_memory %d %d 1" % (self.output_width, self.output_height)
-                            self.agent_process = subprocess.Popen(agent_cmd, shell=True)
+                            subprocess.Popen("pkill -f show_ffmpeg_shared_memory", shell=True)
+                            '''kill_agent_cmd = "sudo -S pkill -f show_ffmpeg_shared_memory"
+                            su_kill_agent_cmd = 'echo {} | '.format(SU_PWD) + kill_agent_cmd
+                            log.debug("su_kill_agent_cmd : %s", kill_agent_cmd)
+                            os.system(su_kill_agent_cmd)'''
+                            time.sleep(1)
+                            ld_path = root_dir + "/ext_binaries"
+                            agent_cmd = (
+                                    "%s/ext_binaries/show_ffmpeg_shared_memory %s %d %d 0 "
+                                    % (root_dir, ETH_DEV, self.output_width, self.output_height))
+                            log.debug("agent_cmd : %s", agent_cmd)
+                            subprocess.Popen(agent_cmd, shell=True)
+                            '''agent_cmd = (
+                                    "LD_LIBRARY_PATH=%s %s/ext_binaries/show_ffmpeg_shared_memory %s %s %d %d 0 &"
+                                    % (ld_path, root_dir, os.getlogin(), ETH_DEV, self.output_width, self.output_height))
+                            # su_agent_cmd = 'echo {} | sudo -S '.format(SU_PWD) + agent_cmd
+                            # os.system(su_agent_cmd)
+                            su_agent_cmd = agent_cmd
+                            os.system(su_agent_cmd)'''
                         continue
                     else:
-                        try_wait_write_flag = 0
-                    
+                        try_wait_write_count = 0
 
                     to_write = memoryview(self.image_from_pipe)
                     self.shm._buf[:len(to_write)] = to_write[:]
-                    #post the read
+                    # post the read
                     read_flag_tmp = self.shm_sem.sem_post(sem_read_flag)
-                    #print("p : ", time.time() - start_time)
-                    #time.sleep(0.001)
-                    # print("e : %s" % time.time())
 
+            log.debug("single play end")
             if self.media_engine.repeat_option == RepeatOption.Repeat_None:
                 log.debug("stop play cause play end")
                 break
@@ -302,8 +322,9 @@ class PlaySingleFileWorker(QObject):
         self.pysignal_single_file_play_finished.emit()
         self.ff_process.kill()
         self.ff_process = None
-        self.agent_process.kill()
-        self.agent_process = None
+        if self.agent_process is not None:
+            self.agent_process.kill()
+            self.agent_process = None
         log.debug("single play worker finished")
 
     def stop(self):
